source/09_GenerativeAI_RPA/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.3.1
openai==1.93.0
packaging==25.0
pandas==2.3.0
pefile==2023.2.7
pydantic==2.11.7
pydantic_core==2.33.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
xlwings==0.33.15
